+++ oci8_go18_test.go
@@ -17,33 +17,6 @@ func TestNamedParam(t *testing.T) {
 	}
 }
 
-/* FIXME
-func TestOutputBind(t *testing.T) {
-	db := DB()
-
-	s1 := "-----------------------------"
-	s2 := 11
-	s3 := false
-	_, err := db.Exec(`begin  :a := 42; :b := 'ddddd' ; :c := 2; end;`,
-		sql.Named("a", sql.Out{Dest: &s2}),
-		sql.Named("b", sql.Out{Dest: &s1}),
-		sql.Named("c", sql.Out{Dest: &s3}))
-	if err != nil {
-		t.Fatal(err)
-	}
-	s1want := "ddddd                        "
-	if s1 != s1want {
-		t.Fatalf("want %q but %q", s1want, s1)
-	}
-	if s2 != 42 {
-		t.Fatalf("want %v but %v", 42, s2)
-	}
-	if !s3 {
-		t.Fatalf("want %v but %v", true, s3)
-	}
-}
-*/
-
 func TestTimeout(t *testing.T) {
 	db := DB()
 	for i := 0; i < 2000; i++ {
